% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crumble.R
\name{crumble}
\alias{crumble}
\title{Mediation Analysis Using Modified Treatment Policies and Recanting Twins}
\usage{
crumble(
  data,
  trt,
  outcome,
  mediators,
  moc,
  covar,
  cens = NULL,
  id = NULL,
  d0 = NULL,
  d1 = NULL,
  learners_regressions = "glm",
  nn_riesz_module = nn_sequential_architecture,
  control = crumble_control()
)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
A \code{data.frame} in wide format containing all necessary variables
for the estimation problem.}

\item{trt}{[\code{character}]\cr
A vector containing the column names of treatment variables.}

\item{outcome}{[\code{character(1)}]\cr
The column name of the outcome variable.}

\item{covar}{[\code{character}]\cr
An optional vector containing the column names of baseline covariates to be
controlled for.}

\item{cens}{[\code{character(1)}]\cr
An optional vector of column name of a censoring indicator. Must be provided if
there is missingness in the outcome.}

\item{id}{[\code{character(1)}]\cr
An optional column name containing cluster level identifiers.}

\item{d0}{[\code{closure}]\cr
A two argument function that specifies how treatment variables should be shifted.
See examples for how to specify shift functions for continuous, binary, and categorical exposures.}

\item{d1}{[\code{closure}\]\cr
A two argument function that specifies how treatment variables should be shifted.
See examples for how to specify shift functions for continuous, binary, and categorical exposures.}

\item{learners_regressions}{[\code{character}]\cr A vector of \code{mlr3superlearner} algorithms
for estimation of the outcome regressions. Default is \code{"glm"}, a main effects GLM.}

\item{control}{}
}
\description{
Mediation Analysis Using Modified Treatment Policies and Recanting Twins
}
