% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_sequential_riesz_representer.R
\name{nn_sequential_architecture}
\alias{nn_sequential_architecture}
\title{Title}
\usage{
nn_sequential_architecture(data, layers = 1, hidden = 20, dropout = 0.1)
}
\arguments{
\item{dropout}{}
}
\description{
Title
}
