% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crumble_control.R
\name{crumble_control}
\alias{crumble_control}
\title{Crumble control parameters}
\usage{
crumble_control(
  crossfit_folds = 10L,
  mlr3superlearner_folds = 10L,
  zprime_folds = 1L,
  epochs = 100L,
  learning_rate = 0.01
)
}
\arguments{
\item{crossfit_folds}{[\code{numeric(1)}]\cr The number of crossfit folds.}

\item{mlr3superlearner_folds}{[\code{numeric(1)}]\cr The number of `mlr3superlearner` folds.}

\item{zprime_folds}{[\code{numeric(1)}]\cr The number of folds to split that data into for calculating Z'.
With larger sample sizes, a larger number will increase speed.}

\item{epochs}{[\code{numeric(1)}]\cr The number of epochs to train the neural network.}

\item{learning_rate}{[\code{numeric(1)}]\cr The learning rate for the neural network.}
}
\value{
A list of control parameters
}
\description{
Crumble control parameters
}
\examples{
crumble_control(crossfit_folds = 5)
}
